Added new React app  using  NX   Failed to process project graph.
An error occurred while processing files for the @nx/eslint/plugin plugin (Defined at nx.json#plugins[2])
.
  - apps/chatbot/eslint.config.js: require is not defined in ES module scope, you can use import instead
This file is being treated as an ES module because it has a '.js' file extension and '/Users/iresta/Developer/learn/my-monorepo/org/package.json' contains "type": "module". To treat it as a CommonJS script, rename it to use the '.cjs' file extension.